top coins

+++ app/src/main/java/com/example/cryptowalletapp/data/retrofit/repository/RetrofitInterface.kt
@@ -5,7 +5,7 @@ import com.example.cryptowalletapp.data.retrofit.model.CoinTop
 import retrofit2.Response
 
 interface RetrofitInterface {
-    suspend fun getCoinHistory(id:String):List<Int>
+    suspend fun getCoinHistory(id:String):List<Double>
     suspend fun getCoinInfo(id:String): CoinDetail
     fun getCoinLogoBySymbol(sym:String):String
     suspend fun getCoinsTopMarket():Response<CoinTop>
